Moved components outside of \pages
