+++ pages/groups/[id].js
@@ -2,7 +2,7 @@ import Head from 'next/head'
 import styles from '../../styles/Home.module.css'
 import postStyles from '../../styles/post.module.css'
 import groupStyles from '../../styles/groups.module.css'
-import Navbar from "../components/Navbar/Navbar"
+import Navbar from "../../components/Navbar/Navbar"
 import { useSession } from 'next-auth/client'
 
 export default function Posts({ posts, group }) {
