+++ pages/groups/new.js
@@ -2,7 +2,7 @@ import Head from 'next/head'
 import styles from '../../styles/Home.module.css'
 import { useState } from 'react'
 import { useRouter } from 'next/router'
-import Navbar from '../components/Navbar/Navbar'
+import Navbar from '../../components/Navbar/Navbar'
 import { useSession } from 'next-auth/client'
 
 const NewGroup = ({ groups }) => {
