+++ pages/index.js
@@ -1,6 +1,6 @@
 import Head from 'next/head'
 import styles from '../styles/Home.module.css'
-import Navbar from "./components/Navbar/Navbar"
+import Navbar from "../components/Navbar/Navbar"
 import { useSession } from 'next-auth/client'
 
 export default function Home() {
